third_party/WebKit/LayoutTests/http/tests/css/pending-stylesheet-offset-width.html

<!DOCTYPE html>
<script src="/js-test-resources/js-test.js"></script>
<style>
    #t { width: 200px; }
</style>
<link rel="stylesheet" href="/resources/load-and-stall.php?name=pending-stylesheet-offset-width.css&stallAt=0&stallFor=1">
<div id="t"></div>
<script>
description("Check that we use the style present asking for offsetWidth, ignoring pending stylesheets.");

shouldBe("t.offsetWidth", "200");
</script>
