third_party/blink/web_tests/platform/generic/http/tests/css/pending-stylesheet-offset-width-expected.txt

Check that we use the style present asking for offsetWidth, ignoring pending stylesheets.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS t.offsetWidth is 200
PASS successfullyParsed is true

TEST COMPLETE
